# Conflicts:
#	gbif/validator/validator-checklists/src/main/java/org/gbif/pipelines/validator/checklists/ChecklistValidator.java
#	livingatlas/pipelines/src/main/java/au/org/ala/pipelines/beam/ALADwcaToVerbatimPipeline.java
#	sdks/beam-transforms/src/main/java/org/gbif/pipelines/factory/FileVocabularyFactory.java
#	sdks/core/src/main/java/org/gbif/pipelines/core/factory/FileSystemFactory.java
#	sdks/core/src/test/java/org/gbif/pipelines/core/converters/GbifJsonConverterTest.java
#	sdks/models/src/main/avro/core/basic-record.avsc

+++ gbif/pipelines/export-gbif-hbase/src/main/java/org/gbif/pipelines/hbase/options/ExportHBaseOptions.java
@@ -18,7 +18,7 @@ public interface ExportHBaseOptions extends HadoopFileSystemOptions {
   void setExportPath(String exportPath);
 
   @Description("Batch size of documents to be read from HBase")
-  @Default.Integer(10000)
+  @Default.Integer(10_000)
   int getBatchSize();
 
   void setBatchSize(int batchSize);
